refactor: S3Client.java 네이밍 변경, 인터페이스 분리, 패키지 분리
