+++ backend/src/main/java/com/mapbefine/mapbefine/image/domain/ImageUploader.java
@@ -1,0 +1,12 @@
+package com.mapbefine.mapbefine.image.domain;
+
+import java.io.IOException;
+import org.springframework.web.multipart.MultipartFile;
+
+public interface ImageUploader {
+
+    void upload(MultipartFile multipartFile) throws IOException;
+
+    void delete(String key);
+
+}
