+++ backend/src/main/java/com/mapbefine/mapbefine/image/infrastructure/S3Uploader.java
@@ -1,8 +1,9 @@
-package com.mapbefine.mapbefine.image.domain;
+package com.mapbefine.mapbefine.image.infrastructure;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.mapbefine.mapbefine.image.domain.ImageUploader;
 import java.io.File;
 import java.io.IOException;
 import java.util.Objects;
@@ -11,16 +12,17 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 @Component
-public class S3Client {
+public class S3Uploader implements ImageUploader {
 
     @Value("${s3.bucket}")
     private String bucket;
     private final AmazonS3 amazonS3;
 
-    public S3Client(AmazonS3 amazonS3) {
+    public S3Uploader(AmazonS3 amazonS3) {
         this.amazonS3 = amazonS3;
     }
 
+    @Override
     public void upload(MultipartFile multipartFile) throws IOException {
         File tempFile = null;
 
@@ -45,8 +47,8 @@ public class S3Client {
         }
     }
 
+    @Override
     public void delete(String key) {
-        // TODO 현재는 일단 기능만 만들어놓고, API 는 만들어놓지 않았습니다 회의를 통해서 결정해야 할 사항이 있는 것 같아서요!
         amazonS3.deleteObject(new DeleteObjectRequest(bucket, key));
     }
